go: Replace application.go with generated file
(split out hand-written extra method to another file)
